Solutions/DataStructuresAndAlgorithms/Hashing/290_WordPattern_test.go
package Solutions

import "testing"

func Test_wordPattern(t *testing.T) {
	type args struct {
		pattern string
		s       string
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		{
			// Input: pattern = "abba", s = "dog cat cat dog"
			// Output: true
			name: "test01",
			args: args{
				pattern: "abba",
				s:       "dog cat cat dog",
			},
			want: true,
		},
		{
			// Input: pattern = "abba", s = "dog cat cat fish"
			// Output: false
			name: "test02",
			args: args{
				pattern: "abba",
				s:       "dog cat cat fish",
			},
			want: false,
		},
		{
			// Input: pattern = "aaaa", s = "dog cat cat dog"
			// Output: false
			name: "test03",
			args: args{
				pattern: "aaaa",
				s:       "dog cat cat dog",
			},
			want: false,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := wordPattern(tt.args.pattern, tt.args.s); got != tt.want {
				t.Errorf("wordPattern() = %v, want %v", got, tt.want)
			}
		})
	}
}
